requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torch-directml>=0.2.0
opencv-python>=4.8.0
albumentations>=1.3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.14.0
